Rename the webservice v1 file to webservice
