+++ src/gateway/api/V1/apartments.py
@@ -26,7 +26,7 @@ from gateway.apartment_controller import ApartmentController
 from gateway.dto import ApartmentDTO
 from gateway.exceptions import WrongInputParametersException, ItemDoesNotExistException
 from gateway.models import User
-from gateway.api.V1.webservice.webservice_v1 import RestAPIEndpoint, openmotics_api_v1, expose
+from gateway.api.V1.webservice.webservice import RestAPIEndpoint, openmotics_api_v1, expose
 
 logger = logging.getLogger(__name__)
