+++ src/gateway/initialize.py
@@ -195,13 +195,13 @@ def setup_target_platform(target_platform, message_client_name):
                          shutter_controller, system_controller, group_action_controller, module_controller,
                          ventilation_controller, apartment_controller, delivery_controller,
                          system_config_controller, rfid_controller, energy_module_controller)
-    from gateway.api.V1.webservice import webservice_v1
+    from gateway.api.V1.webservice import webservice
     from cloud import events
     _ = (metrics_controller, webservice, scheduling_controller, metrics_collector,
          maintenance_controller, base, events, user_controller,
          pulse_counter_controller, metrics_caching, watchdog, output_controller, room_controller,
          sensor_controller, shutter_controller, system_controller, group_action_controller, module_controller,
-         ventilation_controller, webservice_v1, apartment_controller, delivery_controller, system_config_controller,
+         ventilation_controller, webservice, apartment_controller, delivery_controller, system_config_controller,
          rfid_controller, energy_module_controller)
 
     # V1 api
